move lta.py from recon_surf to FastSurferCNN/utils as a more persistent component of FastSurfer (also being used by FastSurferCC)
