--- recon_surf/lta.py
@@ -1,103 +1,0 @@
-#!/usr/bin/env python3
-
-# Copyright 2021 Image Analysis Lab, German Center for Neurodegenerative Diseases (DZNE), Bonn
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from pathlib import Path
-
-import numpy.typing as npt
-
-# Collection of functions related to FreeSurfer's LTA (linear transform array) files:
-
-def write_lta(
-        filename: Path | str,
-        affine: npt.ArrayLike,
-        src_fname: Path | str,
-        src_header: dict,
-        dst_fname: Path | str,
-        dst_header: dict
-) -> None:
-    """
-    Write linear transform array info to an .lta file.
-
-    Parameters
-    ----------
-    filename : Path, str
-        File to write on.
-    affine : npt.ArrayLike
-        Linear transform array to be saved.
-    src_fname : Path, str
-        Source filename.
-    src_header : Dict
-        Source header.
-    dst_fname : Path, str
-        Destination filename.
-    dst_header : Dict
-        Destination header.
-
-    Raises
-    ------
-    ValueError
-        Header format missing field (Source or Destination).
-    """
-    import getpass
-    from datetime import datetime
-
-    fields = ("dims", "delta", "Mdc", "Pxyz_c")
-    for field in fields:
-        if field not in src_header:
-            raise ValueError(f"write_lta Error: src_header format missing field: {field}")
-        if field not in dst_header:
-            raise ValueError(f"write_lta Error: dst_header format missing field: {field}")
-
-    src_dims = str(src_header["dims"][0:3])
-    src_vsize = str(src_header["delta"][0:3])
-    src_v2r = src_header["Mdc"]
-    src_c = src_header["Pxyz_c"]
-
-    dst_dims = str(dst_header["dims"][0:3])
-    dst_vsize = str(dst_header["delta"][0:3])
-    dst_v2r = dst_header["Mdc"]
-    dst_c = dst_header["Pxyz_c"]
-
-    f = open(filename, "w")
-    f.write(
-        (f"# transform file {filename}\n"
-        f"# created by {getpass.getuser()} on {datetime.now().ctime()}\n\n"
-        "type      = 1 # LINEAR_RAS_TO_RAS\n"
-        "nxforms   = 1\n"
-        "mean      = 0.0 0.0 0.0\n"
-        "sigma     = 1.0\n"
-        "1 4 4\n"
-        f"{affine}\n"
-        "src volume info\n"
-        "valid = 1  # volume info valid\n"
-        f"filename = {src_fname}\n"
-        f"volume = {src_dims}\n"
-        f"voxelsize = {src_vsize}\n"
-        f"xras   = {src_v2r[0, :]}\n"
-        f"yras   = {src_v2r[1, :]}\n"
-        f"zras   = {src_v2r[2, :]}\n"
-        f"cras   = {src_c}\n"
-        "dst volume info\n"
-        "valid = 1  # volume info valid\n"
-        f"filename = {dst_fname}\n"
-        f"volume = {dst_dims}\n"
-        f"voxelsize = {dst_vsize}\n"
-        f"xras   = {dst_v2r[0, :]}\n"
-        f"yras   = {dst_v2r[1, :]}\n"
-        f"zras   = {dst_v2r[2, :]}\n"
-        f"cras   = {dst_c}\n").replace("[", "").replace("]", "")
-    )
-    f.close()
